fix(latex): move isTransientHttpStatus to utils, avoid latex->tools edge
Review (Cursor Bugbot, claude[bot]) caught that importing
isTransientHttpStatus from @tools/timeouts into arxivProcessor.ts added
a new latex -> tools subsystem edge absent from
config/ratchets/architecture-edges-baseline.json, on top of the
existing tools -> latex edge — failing the LAY-1 ratchet
(subsystemEdgeRatchet.vitest.ts).

Relocate the pure predicate to src/utils/core/httpStatus.ts, which both
latex and tools already have permitted edges to. timeouts.ts now
imports it for isTransientHttpError; arxivProcessor imports it
directly. Split the isTransientHttpStatus-only unit tests into
test-kernel/utils/HttpStatus.vitest.ts and kept the
isTransientHttpError/isTransientHttpStatus parity check in
TransientHttpError.vitest.ts.
